+++ docs/spec/src/index.ts
@@ -1,7 +1,10 @@
 import { IntegrationSpecConfig } from '@jupiterone/integration-sdk-core';
 
 import { IntegrationConfig } from '../../../src/config';
+import { accountSpec } from './account';
+import { deviceSpec } from './devices';
+import { endUserSpec } from './end-users';
 
 export const invocationConfig: IntegrationSpecConfig<IntegrationConfig> = {
-  integrationSteps: [],
+  integrationSteps: [...accountSpec, ...deviceSpec, ...endUserSpec],
 };